+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Enhancements
 
 * Add support for listing effective tag bindings for a workspace or project by @brandonc
+* Add support for listing no-code modules by @paladin-devops [#1003](https://github.com/hashicorp/go-tfe/pull/1003)
 
 # v1.69.0