+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Unreleased
 
+# v1.70.0
+
 ## Enhancements
 
-* Add support for listing effective tag bindings for a workspace or project by @brandonc
+* Add support for listing effective tag bindings for a workspace or project by @brandonc [#996](https://github.com/hashicorp/go-tfe/pull/996)
 * Add support for listing no-code modules by @paladin-devops [#1003](https://github.com/hashicorp/go-tfe/pull/1003)
 
 # v1.69.0